remove useless grep in quiet refresh
In commit da77c23984bfd3a9ef3a3630ed5eac5393005a0a, getuto "made quiet
mode more quiet". It did this by piping the output of some gpg calls to
a grep pipe, that filtered out any "expected warnings" in the form of no
user IDs emitted by keys.openpgp.org, while mostly leaning on gpg
--quiet to be quiet.

In commit 15fcac948421cf6022a5652d40fbe966dc9aad44, getuto added -q to
grep, with the note "Also, use grep -q to honour --quiet". This meant
that grep was instructed to emit zero output if at all possible, and is
intended to use for scripts that want to check the return code of grep
to see whether matches were found. (getuto does NOT check this.)

Given this is incorrect use of grep, we instead just redirect all stdout
and stderr to /dev/null. It is semantically identical to what we already
do, but with one less fork+exec and the intent is far clearer.